+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Change Log Level on TimeCallback Recalc
+
 ## 2.0.4 (2022-11-24)
 
 * (theimo1221) Add OnChange Log Message to RoomSettingsController